refactor(system/combat): general refactor.
Ahora las explosiones y los spells estan dentro de la carpeta combat. El sistema de particulas esta ahora a la altura de systems
