+++ roguelike_project/systems/combat/explosions/electric.py
@@ -1,6 +1,6 @@
 import random
 import math
-from roguelike_project.systems.combat.view.effects.particles.particle import Particle
+from roguelike_project.systems.particles.particle import Particle
 
 class ElectricExplosion:
     def __init__(self, x, y, particle_count=35):
