+++ roguelike_project/systems/combat/explosions/fire.py
@@ -1,19 +1,21 @@
 import random
 import math
-from roguelike_project.systems.effects.particles.particle import Particle
+from roguelike_project.systems.particles.particle import Particle
 
-class DarkExplosion:
-    def __init__(self, x, y, particle_count=25):
+class FireExplosion:
+    def __init__(self, x, y, particle_count=100):
         self.x = x
         self.y = y
         self.particles = [
             Particle(
                 x, y,
                 angle=random.uniform(0, 2 * math.pi),
-                speed=random.uniform(1, 3),
-                color=random.choice([(40, 0, 40), (60, 0, 60), (20, 20, 20)]),
-                size=random.randint(5, 10),
-                lifespan=random.randint(40, 60)
+                speed=random.uniform(4, 8),
+                color=random.choice([
+                    (255, 80, 0), (255, 180, 0), (255, 255, 0)
+                ]),
+                size=random.randint(6, 10),
+                lifespan=random.randint(20, 35)
             )
             for _ in range(particle_count)
         ]
