+++ roguelike_project/systems/combat/explosions/firework_explosion.py
@@ -2,7 +2,7 @@
 
 import random
 import math
-from roguelike_project.systems.combat.view.effects.particles.particle import Particle
+from roguelike_project.systems.particles.particle import Particle
 
 class FireworkExplosion:
     def __init__(self, x, y, particle_count=60):
