+++ roguelike_project/systems/systems_manager.py
@@ -1,8 +1,8 @@
 # roguelike_project/systems/systems_manager.py
 
-from roguelike_project.systems.combat.view.effects.particles.spells_system import SpellsSystem
+from roguelike_project.systems.combat.spells.spells_system import SpellsSystem
 
-from roguelike_project.systems.combat.view.effects.particles.explosions_system import ExplosionSystem
+from roguelike_project.systems.combat.explosions.explosions_system import ExplosionSystem
 
 class SystemsManager:
     def __init__(self, state):
